+++ Sources/StreamChat/StateLayer/Chat.swift
@@ -1366,6 +1366,7 @@ public class Chat {
     ///   - team: The team for the channel.
     ///   - members: A list of members for the channel.
     ///   - invites: A list of users who will get invites.
+    ///   - filterTags: A list of tags to add to the channel.
     ///   - extraData: Extra data for the new channel.
     ///
     /// - Throws: An error while communicating with the Stream API.
@@ -1375,6 +1376,7 @@ public class Chat {
         team: String?,
         members: Set<UserId> = [],
         invites: Set<UserId> = [],
+        filterTags: Set<String> = [],
         extraData: [String: RawJSON] = [:]
     ) async throws {
         try await channelUpdater.update(
@@ -1385,6 +1387,7 @@ public class Chat {
                 team: team,
                 members: members,
                 invites: invites,
+                filterTags: filterTags,
                 extraData: extraData
             )
         )
@@ -1401,6 +1404,7 @@ public class Chat {
     ///   - team: The team for the channel.
     ///   - members: A list of members for the channel.
     ///   - invites: A list of users who will get invites.
+    ///   - filterTags: A list of tags to add to the channel.
     ///   - extraData: Extra data for the channel.
     ///   - unsetProperties: A list of properties to reset.
     ///
@@ -1411,6 +1415,7 @@ public class Chat {
         team: String? = nil,
         members: [UserId] = [],
         invites: [UserId] = [],
+        filterTags: Set<String> = [],
         extraData: [String: RawJSON] = [:],
         unsetProperties: [String] = []
     ) async throws {
@@ -1422,6 +1427,7 @@ public class Chat {
                 team: team,
                 members: Set(members),
                 invites: Set(invites),
+                filterTags: filterTags,
                 extraData: extraData
             ),
             unsetProperties: unsetProperties